Ignore previous messages from other bots
Only accept previous statements from a human or ourself, and ignore 
those from other bots. This should have the effect of not learning 
messages which are being used as commands to other types of bot.

Changed up the logging a bit to distinguish between "ignoring message 
completely", "not responding to message" and "ignoring message as a 
previous statement".