feat(chatbot): search all vectors using NGT
BREAKING CHANGE: Instead of storing and searching bigram pairs, all statements' vectors are added to an NGT index. The IDs from the index correspond to an entry in the database, which contains nothing but the response text or reaction emoji. This large database-related change means that database files from previous versions will no longer work.