refactor(database): improve database schema
Storing raw messages makes it easier to change the chat algorithm without
having to reset the database, because we can just regenerate the index.

Storing Discord identifiers may be useful for deleting people's data on
request (#36), or for reporting abusive content that is discovered in our
database.

Also, started recording the consent history with timestamps, in case that
comes in useful.

BREAKING CHANGE: The database schema has been significantly changed,
so any previous data will no longer be understood.

Fixes #43