feat: convert relevance into a probability
In general channels, Axyn is now allowed to send less relevant
responses, with an even lower probability than that for a relevant
response.

When directly addressed, Axyn is no longer obliged to reply to
everything, and has a higher chance to not reply when the response would
be less relevant.

Overall, this makes behaviour more consistent between direct and
non-direct channels.