fix: fix error when storing consent prompts / responses
Due to foreign key constraints being enforced now, an error could be
raised when inserting either a `ConsentPromptRecord` or an
`InteractionRecord`, because the related message was not necessarily
inserted yet.

Furthermore, this revealed a deeper problem, which was that we stopped
storing ephemeral messages a while back, and since then there has been
nothing associated with the `ConsentPromptRecord` - so no way to know
which user it relates to.

Fixed this by adding ephemeral messages back to the database, with a
flag that excludes them from being part of the channel history - and
then using `ON CONFLICT DO NOTHING` statements so that we can insert the
message early when we need it without causing an error in the main
handler.